backend/requirements.txt
flask
flask-cors
python-dotenv
requests
groq
google-generativeai
pandas
Pillow
pytesseract
gunicorn
pymupdf
cerebras_cloud_sdk
google-auth-oauthlib
razorpay
firebase-admin

